+++ test/OpenTelemetry.Instrumentation.Http.Tests/HttpWebRequestTests.cs
@@ -19,10 +19,8 @@ public partial class HttpWebRequestTests
 {
     private static readonly JsonSerializerOptions JsonSerializerOptions = new() { PropertyNamingPolicy = JsonNamingPolicy.CamelCase };
 
-    public static IEnumerable<object[]> TestData => HttpTestData.ReadTestCases();
-
     [Theory]
-    [MemberData(nameof(TestData))]
+    [MemberData(nameof(HttpTestData.TestData), MemberType = typeof(HttpTestData))]
     public void HttpOutCallsAreCollectedSuccessfully(HttpOutTestCase tc)
     {
         using var serverLifeTime = TestHttpServer.RunServer(
@@ -151,6 +149,16 @@ public partial class HttpWebRequestTests
 #endif
             Assert.True(enrichWithExceptionCalled);
         }
+
+#if NET
+        // No assertions on Version or TelemetrySchemaUrl as the source is framework-provided
+        Assert.Equal("System.Net.Http", activity.Source.Name);
+#else
+        Assert.Equal("OpenTelemetry.Instrumentation.Http.HttpWebRequest", activity.Source.Name);
+        Assert.NotNull(activity.Source.Version);
+        Assert.NotEmpty(activity.Source.Version);
+        Assert.StartsWith("https://opentelemetry.io/schemas/", activity.Source.TelemetrySchemaUrl);
+#endif
     }
 
     [Fact]
@@ -184,8 +192,6 @@ public partial class HttpWebRequestTests
         this.HttpOutCallsAreCollectedSuccessfully(input);
     }
 
-    private static void ValidateHttpWebRequestActivity(Activity activityToValidate)
-    {
+    private static void ValidateHttpWebRequestActivity(Activity activityToValidate) =>
         Assert.Equal(ActivityKind.Client, activityToValidate.Kind);
-    }
 }